Cleaning up project and final touches
- Changed style for zooming in on website
- Changed visibility styles
- Restructured some code/files
