+++ components/sub-components/typeBadge.tsx
