+++ types/pokemon.ts
@@ -1,0 +1,6 @@
+export type Pokemon = {
+  id: number;
+  name: string;
+  types: string[];
+  sprite: string;
+};
